@@ -4,33 +4,48 @@
   "scripts": {
     "start": "ng serve",
     "docs": "docsify serve docs",
-    "build": "npm run lint && ng build angular8-yandex-maps --prod",
+    "build": "npm run lint && npm run test:ci && ng build angular8-yandex-maps --prod",
+    "test": "ng test angular8-yandex-maps --include=\"**/script.service.spec.ts\"",
+    "test:ci": "ng test angular8-yandex-maps --include=\"**/script.service.spec.ts\" --watch=false --browsers=ChromeHeadless",
     "lint": "ng lint angular8-yandex-maps"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.1.5",
-    "@angular/common": "^10.1.5",
-    "@angular/compiler": "^10.1.5",
-    "@angular/core": "^10.1.5",
-    "@angular/forms": "^10.1.5",
-    "@angular/platform-browser": "^10.1.5",
-    "@angular/platform-browser-dynamic": "^10.1.5",
-    "@angular/router": "^10.1.5",
+    "@angular/animations": "^10.2.0",
+    "@angular/common": "^10.2.0",
+    "@angular/compiler": "^10.2.0",
+    "@angular/core": "^10.2.0",
+    "@angular/forms": "^10.2.0",
+    "@angular/platform-browser": "^10.2.0",
+    "@angular/platform-browser-dynamic": "^10.2.0",
+    "@angular/router": "^10.2.0",
     "rxjs": "~6.5.5",
     "tslib": "^2.0.3",
     "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1000.8",
+    "@angular-devkit/build-angular": "^0.1002.0",
     "@angular-devkit/build-ng-packagr": "^0.1000.8",
-    "@angular/cli": "^10.1.6",
-    "@angular/compiler-cli": "^10.1.5",
-    "@types/jasmine": "^3.5.14",
+    "@angular-eslint/builder": "0.5.0-beta.5",
+    "@angular-eslint/eslint-plugin": "0.5.0-beta.5",
+    "@angular-eslint/eslint-plugin-template": "0.5.0-beta.5",
+    "@angular-eslint/schematics": "^0.5.0-beta.5",
+    "@angular-eslint/template-parser": "0.5.0-beta.5",
+    "@angular/cli": "^10.2.0",
+    "@angular/compiler-cli": "^10.2.0",
+    "@types/jasmine": "^3.6.0",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^12.12.67",
+    "@types/node": "^12.19.1",
+    "@types/yandex-maps": "^2.1.12",
+    "@typescript-eslint/eslint-plugin": "4.3.0",
+    "@typescript-eslint/parser": "4.3.0",
     "codelyzer": "^6.0.1",
     "docsify-cli": "^4.4.1",
+    "eslint": "^7.6.0",
+    "eslint-config-airbnb-typescript": "^12.0.0",
+    "eslint-config-prettier": "^6.14.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-prettier": "^3.1.4",
     "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~5.0.0",
     "karma": "~5.0.0",
@@ -39,10 +54,9 @@
     "karma-jasmine": "~3.3.0",
     "karma-jasmine-html-reporter": "^1.5.0",
     "ng-packagr": "^10.1.2",
+    "prettier": "^2.1.2",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "tslint-config-airbnb": "^5.11.1",
     "typescript": "~3.9.5"
   }
 }
