@@ -8,165 +8,192 @@ import {
   OnDestroy,
   Output,
   QueryList,
-  SimpleChanges
-  } from '@angular/core';
-import { IEvent, ILoadEvent } from '../../models/models';
+  SimpleChanges,
+} from '@angular/core';
 import { startWith } from 'rxjs/operators';
 import { Subscription } from 'rxjs';
+import { IEvent, ILoadEvent } from '../../models/models';
 import { YaGeoObjectComponent } from '../ya-geoobject/ya-geoobject.component';
 import { YaPlacemarkComponent } from '../ya-placemark/ya-placemark.component';
 
 /**
- * Component, geo object clusterer. Clusterizes objects in the visible area of the map.
+ * Component for creating a clusterer. Clusterizes objects in the visible area of the map.
  * If the object does not fall within the visible area of the map, it will not be added to the map.
  * Note, that the clusterer does not react to changing the coordinates of objects (either programmatically,
  * or as the result of dragging). If you want to change the coordinates of some object in the clusterer,
- * you should first delete the object from the clusterer and then add it back
+ * you should first delete the object from the clusterer and then add it back.
  *
  * @example
- * <ya-clusterer>
+ * `<ya-clusterer>
  *  <ya-placemark [geometry]="[55.74, 37.50]"></ya-placemark>
  *   <ya-geoobject [feature]="{ geometry: { type: 'Point', coordinates: [55.73, 37.52] } }"></ya-geoobject>
- * </ya-clusterer>
+ * </ya-clusterer>`.
  * @see {@link https://ddubrava.github.io/angular8-yandex-maps/#/components/clusterer}
  */
 @Component({
   selector: 'ya-clusterer',
   templateUrl: './ya-clusterer.component.html',
-  styleUrls: ['./ya-clusterer.component.scss']
+  styleUrls: ['./ya-clusterer.component.scss'],
 })
 export class YaClustererComponent implements OnDestroy, OnChanges {
   @ContentChildren(YaPlacemarkComponent) public placemarks: QueryList<YaPlacemarkComponent>;
   @ContentChildren(YaGeoObjectComponent) public geoObjects: QueryList<YaGeoObjectComponent>;
 
   /**
-   * Options for the clusterer
-   * @see {@link https://tech.yandex.com/maps/jsapi/doc/2.1/ref/reference/Clusterer-docpage/#Clusterer__param-options}
+   * Options for the clusterer.
+   * @see {@link https://tech.yandex.com/maps/jsapi/doc/2.1/ref/reference/Clusterer-docpage/#Clustererparam-options}
    */
   @Input() public options: any;
 
   /**
-   * Emits immediately after this entity is added in root container
+   * Emits immediately after this entity is added in root container.
    */
   @Output() public load = new EventEmitter<ILoadEvent>();
   /**
-   * Action with hint
+   * Actions with the hint.
    */
   @Output() public hint = new EventEmitter<IEvent>();
   /**
-   * Map reference changed
+   * Map reference changed.
    */
   @Output() public mapChange = new EventEmitter<IEvent>();
   /**
-   * Change to the object options
+   * Change to the object options.
    */
   @Output() public optionsChange = new EventEmitter<IEvent>();
   /**
-   * The parent object reference changed
+   * The parent object reference changed.
    */
   @Output() public parentChange = new EventEmitter<IEvent>();
 
-  private _sub = new Subscription();
+  private sub = new Subscription();
 
-  // Yandex.Maps API
-  private _clusterer: any;
+  // Yandex.Maps API.
+  private clusterer: ymaps.Clusterer;
 
-  constructor(
-    private _ngZone: NgZone,
-  ) { }
+  constructor(private ngZone: NgZone) {}
 
   public ngOnChanges(changes: SimpleChanges): void {
-    this._configClusterer(changes);
+    this.updateClusterer(changes);
   }
 
   /**
-   * Method for dynamic entity configuration.
+   * Method for dynamic Clusterer configuration.
    * Handles input changes and provides it to API.
    * @param changes
    */
-  private _configClusterer(changes: SimpleChanges): void {
-    const clusterer = this._clusterer;
+  private updateClusterer(changes: SimpleChanges): void {
+    const { clusterer } = this;
 
     if (!clusterer) return;
 
     const { options } = changes;
 
     if (options) {
-      clusterer.options.set(options.currentValue);
+      /**
+       * Wrong typings in DefinitelyTyped.
+       */
+      (clusterer.options as any).set(options.currentValue);
     }
   }
 
-  public initClusterer(ymaps: any, map: any): void {
+  /**
+   * Creates Clusterer.
+   *
+   * @param map Necessary for passing in children.
+   */
+  public createClusterer(map: ymaps.Map): ymaps.Clusterer {
     const clusterer = new ymaps.Clusterer(this.options);
-    this._clusterer = clusterer;
+    this.clusterer = clusterer;
 
     /**
-     * Adds new Placemarks to clusterer on changes
+     * Adds new Placemarks to the clusterer on changes.
      */
     const placemarksSub = this.placemarks.changes
       .pipe(startWith(this.placemarks))
       .subscribe((list: QueryList<YaPlacemarkComponent>) => {
         list.forEach((placemark: YaPlacemarkComponent) => {
           if (!placemark.id) {
-            clusterer.add(placemark.initPlacemark(ymaps, map, clusterer));
+            /**
+             * Wrong typings in DefinitelyTyped.
+             */
+            (clusterer as any).add(placemark.createPlacemark(map, clusterer));
           }
         });
       });
 
-    this._sub.add(placemarksSub);
+    this.sub.add(placemarksSub);
 
     /**
-     * Adds new GeoObjects to clusterer on changes
+     * Adds new GeoObjects to the clusterer on changes.
      */
     const geoObjectsSub = this.geoObjects.changes
       .pipe(startWith(this.geoObjects))
       .subscribe((list: QueryList<YaGeoObjectComponent>) => {
         list.forEach((geoObject: YaGeoObjectComponent) => {
           if (!geoObject.id) {
-            clusterer.add(geoObject.initGeoObject(ymaps, map, clusterer));
+            /**
+             * Wrong typings in DefinitelyTyped.
+             */
+            (clusterer as any).add(geoObject.createGeoObject(map, clusterer));
           }
         });
       });
 
-    this._sub.add(geoObjectsSub);
+    this.sub.add(geoObjectsSub);
 
-    map.geoObjects.add(clusterer);
-    this._addEventListeners(ymaps, clusterer);
+    this.addEventListeners();
+
+    return clusterer;
   }
 
   /**
-   * Add listeners on clusterer events
-   * @param ymaps
-   * @param map
+   * Adds listeners on Clusterer events.
    */
-  private _addEventListeners(ymaps: any, clusterer: any): void {
-    this._ngZone.run(() => this.load.emit({ ymaps, instance: clusterer }));
+  private addEventListeners(): void {
+    const { clusterer } = this;
+
+    this.ngZone.run(() => this.load.emit({ ymaps, instance: clusterer }));
 
     const handlers = [
       {
         name: ['hintclose', 'hintopen'],
-        fn: (e: any) => this.hint.emit({ ymaps, instance: clusterer, type: e.originalEvent.type, event: e }),
+        fn: (e: any) =>
+          this.hint.emit({ ymaps, instance: clusterer, type: e.originalEvent.type, event: e }),
       },
       {
         name: 'mapchange',
-        fn: (e: any) => this.mapChange.emit({ ymaps, instance: clusterer, type: e.originalEvent.type, event: e }),
+        fn: (e: any) =>
+          this.mapChange.emit({ ymaps, instance: clusterer, type: e.originalEvent.type, event: e }),
       },
       {
         name: 'optionschange',
-        fn: (e: any) => this.optionsChange.emit({ ymaps, instance: clusterer, type: e.originalEvent.type, event: e }),
+        fn: (e: any) =>
+          this.optionsChange.emit({
+            ymaps,
+            instance: clusterer,
+            type: e.originalEvent.type,
+            event: e,
+          }),
       },
       {
         name: 'parentchange',
-        fn: (e: any) => this.parentChange.emit({ ymaps, instance: clusterer, type: e.originalEvent.type, event: e }),
+        fn: (e: any) =>
+          this.parentChange.emit({
+            ymaps,
+            instance: clusterer,
+            type: e.originalEvent.type,
+            event: e,
+          }),
       },
     ];
 
     handlers.forEach((handler) => {
-      clusterer.events.add(handler.name, (e: any) => this._ngZone.run(() => handler.fn(e)));
+      clusterer.events.add(handler.name, (e: any) => this.ngZone.run(() => handler.fn(e)));
     });
   }
 
   public ngOnDestroy(): void {
-    this._sub.unsubscribe();
+    this.sub.unsubscribe();
   }
 }
