@@ -3,55 +3,59 @@ import {
   EventEmitter,
   Input,
   NgZone,
+  OnChanges,
   OnInit,
   Output,
-  SimpleChanges
-  } from '@angular/core';
+  SimpleChanges,
+} from '@angular/core';
 import { ILoadEvent } from '../../models/models';
 import { removeLeadingSpaces } from '../../utils/removeLeadingSpaces';
 
 /**
- * Component for creating and managing controls on the map
- * @example <ya-control type="RoutePanel" [parameters]="{ options: { float: 'right' } }"></ya-control>
+ * Component for creating and managing controls on the map.
+ *
+ * @example `<ya-control type="RoutePanel" [parameters]="{ options: { float: 'right' } }"></ya-control>`.
  * @see {@link https://ddubrava.github.io/angular8-yandex-maps/#/components/controls}
  */
 @Component({
   selector: 'ya-control',
   templateUrl: './ya-control.component.html',
-  styleUrls: ['./ya-control.component.scss']
+  styleUrls: ['./ya-control.component.scss'],
 })
-export class YaControlComponent implements OnInit {
+export class YaControlComponent implements OnInit, OnChanges {
   /**
-   * Control type
-   * @example Control.FullscreenControl - 'FullscreenControl'
+   * Control type.
+   * @example Control.FullscreenControl - 'FullscreenControl'.
    * @see {@link https://tech.yandex.ru/maps/jsapi/doc/2.1/ref/reference/control.Button-docpage/}
    */
   @Input() public type: string;
   /**
-   * Parameters for the Control
+   * Parameters for the Control.
    */
   @Input() public parameters: any;
 
   /**
-   * Emits immediately after this entity is added in root container
+   * Emits immediately after this entity is added in root container.
    */
   @Output() public load = new EventEmitter<ILoadEvent>();
 
-  constructor(
-    private _ngZone: NgZone,
-  ) {}
+  constructor(private ngZone: NgZone) {}
 
   public ngOnInit(): void {
-    this._logErrors();
+    this.logErrors();
   }
 
-  private _logErrors(): void {
+  private logErrors(): void {
     if (!this.type) {
       console.error('Control: type input is required.');
     }
   }
 
-  public initControl(ymaps: any, map: any): void {
+  /**
+   * Creates control
+   * @returns Instance of created control
+   */
+  public createControl(): any {
     const control = new ymaps.control[this.type](this.parameters);
 
     // RoutePanel ignores state in parameters. API bug
@@ -59,30 +63,30 @@ export class YaControlComponent implements OnInit {
       control.routePanel.state.set({ ...this.parameters.state });
     }
 
-    map.controls.add(control);
+    this.emitEvent(control);
 
-    this._emitEvent(ymaps, control);
+    return control;
   }
 
   public ngOnChanges(changes: SimpleChanges): void {
-    this._configControl(changes);
+    this.updateControl(changes);
   }
 
-  private _configControl(changes: SimpleChanges): void {
-    for (const key in changes) {
-      if (changes[key].firstChange) return;
-    }
+  private updateControl(changes: SimpleChanges): void {
+    if (Object.values(changes).some((v) => v.firstChange)) return;
 
-    console.error(removeLeadingSpaces(`
+    console.error(
+      removeLeadingSpaces(`
       Control doesn't support dynamic configuartion.
 
       Solutions:
       1. Use ymaps from ILoadEvent
       2. Recreate component with new configuration
-    `));
+    `),
+    );
   }
 
-  private _emitEvent(ymaps: any, control: any): void {
-    this._ngZone.run(() => this.load.emit({ ymaps, instance: control }));
+  private emitEvent(control: any): void {
+    this.ngZone.run(() => this.load.emit({ ymaps, instance: control }));
   }
 }
