@@ -3,6 +3,6 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'app-clusterer',
   templateUrl: './clusterer.component.html',
-  styleUrls: ['./clusterer.component.css']
+  styleUrls: ['./clusterer.component.css'],
 })
 export class ClustererComponent {}
