@@ -3,6 +3,6 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'app-geoobject',
   templateUrl: './geoobject.component.html',
-  styleUrls: ['./geoobject.component.css']
+  styleUrls: ['./geoobject.component.css'],
 })
 export class GeoobjectComponent {}
