@@ -3,6 +3,6 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'app-multiroute',
   templateUrl: './multiroute.component.html',
-  styleUrls: ['./multiroute.component.css']
+  styleUrls: ['./multiroute.component.css'],
 })
 export class MultirouteComponent {}
