@@ -3,6 +3,6 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'app-panorama',
   templateUrl: './panorama.component.html',
-  styleUrls: ['./panorama.component.css']
+  styleUrls: ['./panorama.component.css'],
 })
 export class PanoramaComponent {}
