@@ -3,6 +3,6 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'app-placemark',
   templateUrl: './placemark.component.html',
-  styleUrls: ['./placemark.component.css']
+  styleUrls: ['./placemark.component.css'],
 })
 export class PlacemarkComponent {}
